updated files
update added
Remove the "(Optional)" tag
LinkedIn link in the Readme
Fixed issue with the vertical scroll
Replacing the <section> tag with an <article> tag for the card element
Remove the gap between the card
updated section subtitle!
proper spacing is added around the button.
alignment of the button text
Review the font styles of the button, including font size
proper spacing is added between the social icons
included <h2>, for the section subtitle instead of a <p>